requirements.txt
aiohttp==3.8.5
discord.py==2.3.2
python-dotenv==1.0.0
cachetools==5.3.1
fake-useragent==1.2.0
Pillow==10.0.1
scikit-learn==1.7.1